requirements.txt
flask
loguru
pyYAML